+++ .prettierrc.json
@@ -1,7 +1,7 @@
 {
   "semi": false,
   "tabWidth": 2,
-  "printWidth": 110,
+  "printWidth": 125,
   "singleQuote": true,
   "trailingComma": "all",
   "jsxSingleQuote": true,
