+++ package.json
@@ -17,6 +17,7 @@
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
     "recharts": "^2.1.14",
+    "sass": "^1.55.0",
     "typescript": "^4.8.2",
     "web-vitals": "^2.1.4"
   },
