+++ src/assets/images/activityItems/caloriesIcon.png
