+++ src/assets/images/activityItems/carbsIcon.png
