+++ src/assets/images/activityItems/proteinIcon.png
