+++ src/assets/images/sideBar/swimmingButton.png
