+++ src/assets/images/sideBar/yogaButton.png
