+++ src/components/charts/activityItems/_activity.scss
