+++ src/components/charts/barChart/_barChart.scss
@@ -1,3 +1,8 @@
+.barChart-svg {
+  background-color: #f5f7f9;
+  border-radius: 5px;
+}
+
 .bar_chart {
   grid-area: 2 / 1 / 3 / 4;
 }
