+++ src/components/charts/radialChart/_radialChart.scss
@@ -1,3 +1,7 @@
+.radial-chart-svg {
+  background-color: #f5f7f9;
+  border-radius: 5px;
+}
 .radial_chart {
     grid-area: 1 / 3 / 2 / 4; 
 }
