+++ src/components/charts/spiderChart/_spiderChart.scss
