+++ src/components/static/sideBar/sidebar.scss
@@ -14,7 +14,7 @@
   }
   // Footer Text
   &_text {
-    color: $back-white;
+    color: white;
     font-size: 12px;
     transform: rotate(-90deg);
     white-space: nowrap;
