+++ src/components/static/topMessage/topMessage.scss
