--- src/sass/modules/_collapse.scss
@@ -1,54 +1,0 @@
-/**************** Collapse Element ****************/
-.collapse {
-  @include wh-100;
-  border-radius: 10px;
-  // Collapse Title
-  &_title {
-    position: relative;
-    background-color: $primary-color;
-    width: 100%;
-    padding: 10px 25px 10px 25px;
-    border-radius: 10px;
-    font-size: 18px;
-    font-weight: 500;
-    color: white;
-    display: flex;
-    justify-content: space-between;
-    // Arrow unrotate to initial position
-    & .unrotate-arrow {
-      transform: rotate(0deg);
-      transition: 0.5s ease-in-out;
-    }
-    // Arrow rotation
-    & .rotate-arrow {
-      transform: rotate(-180deg);
-      transition: 0.5s ease-in-out;
-    }
-  }
-  // Collapse Content
-  &_content {
-    width: 100%;
-    padding: 45px 25px 30px 25px;
-    margin-top: -10px;
-    font-size: 18px;
-    font-weight: 400;
-    color: $primary-color;
-    background-color: $light-grey;
-    border-radius: 10px;
-  }
-}
-/**************** Responsive ****************/
-@media screen and (max-width: 768px) {
-  .collapse {
-    &_title {
-      font-size: 13px;
-      & .rotate-arrow {
-        width: 18px;
-      }
-    }
-    &_content {
-      padding: 20px 10px 10px 10px;
-      font-size: 12px;
-    }
-  }
-}
