--- src/sass/modules/_lightbox.scss
@@ -1,63 +1,0 @@
-/**************** Lightbox ****************/
-.lightbox {
-  @include flex-row-center;
-  position: relative;
-  margin-bottom: 30px;
-  mix-blend-mode: normal;
-  // Lightbox Image
-  &_image {
-    @include wh(85vw, 45.5vh);
-    -o-object-fit: cover;
-    object-fit: cover;
-    border-radius: 25px;
-  }
-  // Lightbox Index
-  &_index {
-    position: absolute;
-    bottom: 30px;
-    left: 50%;
-    color: $back-white;
-    font-size: 18px;
-    font-weight: 500;
-    text-shadow: black 1px 1px;
-  }
-  // Chevron for next and previous image
-  .chevron {
-    @include wh(47px, 70px);
-    position: absolute;
-    fill: $back-white;
-    cursor: pointer;
-    stroke: rgba(0, 0, 0, 0.507);
-    stroke-width: 10;
-  }
-  // floating left position
-  .left {
-    left: 70px;
-  }
-  // floating right position
-  .right {
-    right: 70px;
-  }
-}
-/**************** Responsive ****************/
-@media screen and (max-width: 768px) {
-  .lightbox {
-    &_image {
-      height: 30vh;
-      border-radius: 10px;
-      min-width: 320px;
-    }
-    &_index {
-      display: none;
-    }
-    .chevron {
-      @include wh(20px, 30px);
-    }
-    .left {
-      left: 10px;
-    }
-    .right {
-      right: 10px;
-    }
-  }
-}
