+++ src/sass/utils/variables.scss
@@ -7,5 +7,4 @@ $light-background: #fbfbfb;
 $tooltip-hover: grey;
 $sessions-graph-legend: #9b9eac;
 $sessions-graph-colomn: #c4c4c4;
-
 $back-white: white;
