--- src/utils/d3Generics.ts
@@ -1,9 +1,0 @@
-export const groups = (
-  type: string,
-  x: number,
-  y: number,
-  width: string | number,
-  height: string | number,
-  className?: string,
-  text?: string,
-) => {}
