+++ src/utils/responsive.ts
@@ -1,4 +1,3 @@
-
 export const vh = (percent) => {
   const screen = Math.max(document.documentElement.clientHeight, window.innerHeight || 0),
     h = (percent * screen) / 100
@@ -11,5 +10,20 @@ export const vw = (percent) => {
   return w > 155 ? w : 154
 }
 
+export const svgRadius = (width) => {
+  let radius = 50,
+    oRadius = 55,
+    iRadius = 48
 
+  if (width > 248) {
+    radius = 90
+    oRadius = 100
+    iRadius = 80
+  } else if (width > 200 && width <= 248) {
+    radius = 70
+    oRadius = 75
+    iRadius = 65
+  }
 
+  return { radius, oRadius, iRadius }
+}
